+++ src/components/Header/index.tsx
@@ -1,0 +1,4 @@
+const Header = () => {
+  return <h1 className="h1">{`Wookie Movies`}</h1>;
+};
+export default Header;