+++ src/pages/Movies/index.tsx
@@ -1,16 +1,8 @@
 // src/MoviesByGenre.tsx
 import React from "react";
 import { useQuery } from "@tanstack/react-query";
-import { fetchMoviesByGenre, fetchGenres } from "../../apis/moviesApi";
-import MovieCard from "../../components/MovieCard.tsx";
-
-interface Movie {
-  id: number;
-  title: string;
-  overview: string;
-  poster_path: string;
-  release_date: string;
-}
+import { fetchGenres } from "../../apis/moviesApi";
+import GenreSection from "../../components/GenreSection/index.tsx";
 
 interface Genre {
   id: number;
@@ -27,42 +19,20 @@ const MoviesByGenre: React.FC = () => {
     queryFn: fetchGenres,
   });
 
-  if (genresLoading) return <div>Loading genres...</div>;
-  if (genresError) return <div>Error: {genresError.message}</div>;
+  if (genresLoading)
+    return <div className="text-center">Loading genres...</div>;
+  if (genresError)
+    return (
+      <div className="text-center text-red-500">
+        Error: {genresError.message}
+      </div>
+    );
 
   return (
-    <div>
+    <div className="space-y-8">
       {genres?.genres.map((genre: Genre) => (
         <GenreSection key={genre.id} genre={genre} />
       ))}
-    </div>
-  );
-};
-
-const GenreSection: React.FC<{ genre: Genre }> = ({ genre }) => {
-  const { data, error, isLoading } = useQuery({
-    queryKey: ["moviesByGenre", genre.id],
-    queryFn: () => fetchMoviesByGenre(genre.id),
-  });
-
-  if (isLoading) return <div>Loading movies for {genre.name}...</div>;
-  if (error) return <div>Error: {error.message}</div>;
-
-  return (
-    <div>
-      <h2>{genre.name}</h2>
-      <div style={{ display: "flex", flexWrap: "wrap", gap: "16px" }}>
-        {data?.results.map((movie: Movie) => (
-         <MovieCard
-         key={movie.id}
-         id={movie.id}
-         title={movie.title}
-         poster_path={movie.poster_path}
-         release_date={movie.release_date}
-         overview={movie.overview}
-       />
-        ))}
-      </div>
     </div>
   );
 };